+++ lib/active_record_compose/inner_model.rb
@@ -60,12 +60,36 @@ module ActiveRecordCompose
     # Whether save or destroy is executed depends on the value of `#destroy_context?`.
     #
     # @return [Boolean] returns true on success, false on failure.
-    def save = destroy_context? ? model.destroy : model.save
+    def save
+      # While errors caused by the type check are avoided,
+      # it is important to note that an error can still occur
+      # if `#destroy_context?` returns true but ar_like does not implement `#destroy`.
+      m = model
+      if destroy_context?
+        # @type var m: ActiveRecordCompose::_ARLikeWithDestroy
+        m.destroy
+      else
+        # @type var m: ActiveRecordCompose::_ARLike
+        m.save
+      end
+    end
 
     # Execute save or destroy. Unlike #save, an exception is raises on failure.
     # Whether save or destroy is executed depends on the value of `#destroy_context?`.
     #
-    def save! = destroy_context? ? model.destroy! : model.save!
+    def save!
+      # While errors caused by the type check are avoided,
+      # it is important to note that an error can still occur
+      # if `#destroy_context?` returns true but ar_like does not implement `#destroy`.
+      m = model
+      if destroy_context?
+        # @type var m: ActiveRecordCompose::_ARLikeWithDestroy
+        m.destroy!
+      else
+        # @type var model: ActiveRecordCompose::_ARLike
+        m.save!
+      end
+    end
 
     # @return [Boolean]
     def invalid? = destroy_context? ? false : model.invalid?